pyproject.toml
[project]
name = "car-plate-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dvc>=3.66.1",
    "opencv-python>=4.8.0",
    "pytesseract>=0.3.13",
    "ultralytics>=8.4.21",
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0,<2",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
]
